separate some utility functions out into separate files
