+++ pyveg/src/date_utils.py
@@ -1,0 +1,80 @@
+"""
+Useful functions for manipulating dates and date strings,
+e.g. splitting a period into sub-periods.
+
+When dealing with date strings, ALWAYS use the ISO format YYYY-MM-DD
+
+"""
+
+import dateparser
+from datetime import datetime, timedelta
+
+
+def get_num_n_day_slices(start_date, end_date, days_per_chunk):
+    """
+    Divide the full period between the start_date and end_date into n equal-length
+    (to the nearest day) chunks. The size of the chunk is defined by days_per_chunk.
+    Takes start_date and end_date as strings 'YYYY-MM-DD'.
+    Returns an integer with the number of possible points avalaible in that time period]
+    """
+    start = dateparser.parse(start_date)
+    end = dateparser.parse(end_date)
+    if (not isinstance(start, datetime)) or (not isinstance(end, datetime)):
+        raise RuntimeError("invalid time strings")
+    td = end - start
+    if td.days <= 0:
+        raise RuntimeError("end_date must be after start_date")
+    n = td.days//days_per_chunk
+
+    return  n
+
+
+def find_mid_period(start_time, end_time):
+    """
+    Given two strings in the format YYYY-MM-DD return a
+    string in the same format representing the middle (to
+    the nearest day)
+
+    Parameters
+    ==========
+    start_time: str, date in format YYYY-MM-DD
+    end_time: str, date in format YYYY-MM-DD
+
+    Returns
+    =======
+    mid_date: str, mid point of those dates, format YYYY-MM-DD
+    """
+    t0 = dateparser.parse(start_time)
+    t1 = dateparser.parse(end_time)
+    td = (t1 - t0).days
+    mid = (t0 + timedelta(days=(td//2))).isoformat()
+    mid_date = mid.split("T")[0]
+    return mid_date
+
+
+def slice_time_period(start_date, end_date, n):
+    """
+    Divide the full period between the start_date and end_date into n equal-length
+    (to the nearest day) chunks.
+    Takes start_date and end_date as strings 'YYYY-MM-DD'.
+    Returns a list of tuples
+    [ (chunk0_start,chunk0_end),...]
+    """
+    start = dateparser.parse(start_date)
+    end = dateparser.parse(end_date)
+    if (not isinstance(start, datetime)) or (not isinstance(end, datetime)):
+        raise RuntimeError("invalid time strings")
+    td = end - start
+    if td.days <= 0:
+        raise RuntimeError("end_date must be after start_date")
+    days_per_chunk = td.days // n
+    output_list = []
+    for i in range(n):
+        chunk_start = start + timedelta(days=(i*days_per_chunk))
+        chunk_end = start + timedelta(days=((i+1)*days_per_chunk))
+        ## unless we are in the last chunk, which should finish at end_date
+        if i == n-1:
+            chunk_end = end
+        output_list.append((chunk_start.isoformat().split("T")[0],
+                           chunk_end.isoformat().split("T")[0]))
+    return output_list
