+++ pyveg/src/file_utils.py
@@ -19,65 +19,6 @@ def find_coords_string(file_path):
     return coords_string
 
 
-def find_mid_period(start_time, end_time):
-    """
-    Given two strings in the format YYYY-MM-DD return a
-    string in the same format representing the middle (to
-    the nearest day)
-    """
-    t0 = dateparser.parse(start_time)
-    t1 = dateparser.parse(end_time)
-    td = (t1 - t0).days
-    mid = (t0 + timedelta(days=(td//2))).isoformat()
-    return mid.split("T")[0]
-
-
-def slice_time_period(start_date, end_date, n):
-    """
-    Divide the full period between the start_date and end_date into n equal-length
-    (to the nearest day) chunks.
-    Takes start_date and end_date as strings 'YYYY-MM-DD'.
-    Returns a list of tuples
-    [ (chunk0_start,chunk0_end),...]
-    """
-    start = dateparser.parse(start_date)
-    end = dateparser.parse(end_date)
-    if (not isinstance(start, datetime)) or (not isinstance(end, datetime)):
-        raise RuntimeError("invalid time strings")
-    td = end - start
-    if td.days <= 0:
-        raise RuntimeError("end_date must be after start_date")
-    days_per_chunk = td.days // n
-    output_list = []
-    for i in range(n):
-        chunk_start = start + timedelta(days=(i*days_per_chunk))
-        chunk_end = start + timedelta(days=((i+1)*days_per_chunk))
-        ## unless we are in the last chunk, which should finish at end_date
-        if i == n-1:
-            chunk_end = end
-        output_list.append((chunk_start.isoformat().split("T")[0],
-                           chunk_end.isoformat().split("T")[0]))
-    return output_list
-
-
-def get_num_n_day_slices(start_date, end_date, days_per_chunk):
-    """
-    Divide the full period between the start_date and end_date into n equal-length
-    (to the nearest day) chunks. The size of the chunk is defined by days_per_chunk.
-    Takes start_date and end_date as strings 'YYYY-MM-DD'.
-    Returns an integer with the number of possible points avalaible in that time period]
-    """
-    start = dateparser.parse(start_date)
-    end = dateparser.parse(end_date)
-    if (not isinstance(start, datetime)) or (not isinstance(end, datetime)):
-        raise RuntimeError("invalid time strings")
-    td = end - start
-    if td.days <= 0:
-        raise RuntimeError("end_date must be after start_date")
-    n = td.days//days_per_chunk
-
-    return  n
-
 
 def save_json(out_dict, output_dir, output_filename):
     """
@@ -92,3 +33,41 @@ def save_json(out_dict, output_dir, output_filename):
         json.dump(out_dict, fp, indent=2)
 
     print("Saved json file '{}'".format(output_path))
+
+
+def construct_image_savepath(output_dir, collection_name, coords, date_range, image_type):
+    """
+    Function to abstract output image filename construction. Current approach is to create
+    a new dir inside `output_dir` for the satellite, and then save date and coordinate
+    stamped images in this dir.
+    """
+
+    # get the mid point of the date range
+    mid_period_string = find_mid_period(date_range[0], date_range[1])
+
+    # filename is the date, coordinates, and image type
+    filename = f'{mid_period_string}_{coords[0]}-{coords[1]}_{image_type}.png'
+
+    # full path is dir + filename
+    full_path = os.path.join(output_dir, filename)
+
+    return full_path
+
+
+def consolidate_json_to_list(json_dir, output_dir, output_filename="network_centralities.json"):
+    """
+    Load all the json files (e.g. from individual sub-images), and return
+    a list of dictionaries, to be written out into one json file.
+
+    Parameters
+    ==========
+    json_dir: str, full path to directory containing temporary json files
+    output_dir: str, full path to desired output directory
+    output_filename: str, name of the output json file.
+    """
+    results = []
+
+    for filename in os.listdir(json_dir):
+        results.append(json.load(open(os.path.join(json_dir,filename))))
+    save_json(results, output_dir, output_filename)
+    return results
